refactor: enhanced syscall patch compatibility and removed support for kernel version < 3.18 devices (ARMV7A).
Since KernelSU Fork does not always respond to Syscall updates in a
timely manner, it is possible that it is still using an older version of
Syscall, so it has been decided to incorporate and make compatible some
of the 1.5 parts to avoid problems such as signature errors.
In addition, due to the scarcity of users and the substantial increase
in maintenance costs, I have decided to remove support for ARMV7A
(kernel versions 3.4-3.10) devices.

It should still be noted that, for example, KernelSU Next and
KernelSU(rsuntk) include, they need to use the CI version manager rather
than the release version, which suffers from erroneous signature
exception issues.

Changes:
- Incorporate syscall patch 1.5 into version 1.6 and implement
auto-determination.
- Removed syscall_hook_patches_older.sh.
- Removed support for kernel version < 3.18 devices (ARMV7A).
- Enhanced compatibility with current version 1.6 patches.

Co-authored-by: JackAltman
<82563979+JackA1ltman@users.noreply.github.com>